Added backend and db
NOTES FOR RUNNING THIS VERSION:
NOTE: You have to have the backend up and running, otherwise the app crashes after the login page (not a bug, it tries to fetch data from the backend/ database).
NOTE: When downloading the file you should change the username and password in the properties file in order for it to connect to MySQL
NOTE: You have to have MySQL Server on your device as that's our db
NOTE: You don't need a db to run the project, it will automatically create one for you. No filled db exists as of now, as the backend logic is not fully developed.

KEY CHANGES OF THIS UPDATE:
UPDATE: Added the database and backend to the android application.
UPDATE: The dashboard now shows non-static data, aka it fetches data from the database.

(CURRENTLY WORKING ON: having user information such as SmartSaveProfile in order to be able to fetch information about a business and be able to calculate and display it)